+++ profiles.clj
@@ -1,0 +1,1 @@
+{:dev {:dependencies [[funcool/promesa "0.6.0"]]}}