+++ CHANGES.md
@@ -7,6 +7,7 @@ Date: Unreleased
 
 - Fix wrong precondition on `repeat` function.
 - Add `scan` function.
+- Add `from-promise` function.
 
 ## Version 0.2.0 ##